+++ android/src/main/java/com/yendoplan/openappfile/FileProvider.java
@@ -1,4 +1,4 @@
-package com.crazecoder.openfile;
+package com.yendoplan.openappfile;
 
 public class FileProvider extends androidx.core.content.FileProvider {
 }