+++ android/src/main/java/com/yendoplan/openappfile/utils/MapUtil.java
@@ -1,4 +1,4 @@
-package com.crazecoder.openfile.utils;
+package com.yendoplan.openappfile.utils;
 
 import java.util.HashMap;
 import java.util.Map;