+++ lib/src/plaform/open_app_file.dart
@@ -3,13 +3,13 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:flutter/services.dart';
-import 'package:open_file/src/common/open_result.dart';
+import 'package:open_app_file/src/common/open_result.dart';
 import 'macos.dart' as mac;
 import 'windows.dart' as windows;
 import 'linux.dart' as linux;
 
 class OpenFile {
-  static const MethodChannel _channel = const MethodChannel('open_file');
+  static const MethodChannel _channel = const MethodChannel('open_app_file');
 
   OpenFile._();
 
@@ -54,7 +54,7 @@ class OpenFile {
       "type": type,
       "uti": uti,
     };
-    final _result = await _channel.invokeMethod('open_file', map);
+    final _result = await _channel.invokeMethod('open_app_file', map);
     final resultMap = json.decode(_result) as Map<String, dynamic>;
     return OpenResult.fromJson(resultMap);
   }